OAK-9132: Feature toggles
Rename classes as suggested by Julian Sedding
